Modified the networking functions to now encode all fields bigger than a byte to big endian representations, then receive and transform them to the host system represntation.

+++ net.h
@@ -14,7 +14,7 @@ int send_protocol_message(int fd,
     uint8_t msg_type,
     uint8_t sender_id,
     uint8_t target_id,
-    size_t payload_len,
+    uint16_t payload_len,
     const void *payload);
 
 DECL_SEND_FN(hello, msg_hello_t);
